Refactor : 전체 Container 리팩토링 AppRoute에서 컨테이너 규격 지원 Related To:#3
